refactor: :recycle: rename directory to add semanthic clarity
