+++ src/app.module.ts
@@ -7,15 +7,15 @@ import { UsersModule } from './modules/users/users.module';
 import { PhotosModule } from './modules/photos/photos.module';
 import { DatabaseConfigService } from './config/database/database-options.constants';
 import configuration from './config/configuration';
-import { validate } from './resources/validations/env.validation';
+import { validate } from './common/validations/env.validation';
 import { MainModule } from './main.module';
 import { APP_FILTER, APP_INTERCEPTOR } from '@nestjs/core';
-import { HttpExceptionFilter } from './resources/filters/http-exception.filters';
+import { HttpExceptionFilter } from './common/filters/http-exception.filters';
 import { CacheModule } from '@nestjs/cache-manager';
 import { CacheConfig } from './config/modules/cache.config';
-import { HttpCacheInterceptor } from './resources/interceptors/http-cache.interceptor';
+import { HttpCacheInterceptor } from './common/interceptors/http-cache.interceptor';
 import { AuthModule } from './modules/auth/auth.module';
-import { GlobalLoggerInterceptor } from './resources/interceptors/global-logger.interceptor';
+import { GlobalLoggerInterceptor } from './common/interceptors/global-logger.interceptor';
 import { ServeStaticModule } from '@nestjs/serve-static';
 import { join } from 'path';
 import { AIModule } from './modules/ai/ai.module';
