+++ src/modules/auth/auth.guard.ts
@@ -2,7 +2,7 @@ import { CanActivate, ExecutionContext, Injectable, UnauthorizedException } from
 import { Reflector } from "@nestjs/core";
 import { JwtService } from "@nestjs/jwt";
 import { Request } from "express";
-import { IS_PUBLIC_KEY } from "../../resources/decorators/public-route.decorator";
+import { IS_PUBLIC_KEY } from "../../common/decorators/public-route.decorator";
 import { UserPayload } from "./auth.service";
 
 export interface UserRequest extends Request {
