+++ src/modules/users/users.controller.ts
@@ -18,9 +18,9 @@ import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './entities/user.entity';
-import { HashingPassword } from '../../resources/pipes/hashing-password.pipe';
+import { HashingPassword } from '../../common/pipes/hashing-password.pipe';
 import { ListUserDto } from './dto/list-user.dto';
-import { Public } from '../../resources/decorators/public-route.decorator';
+import { Public } from '../../common/decorators/public-route.decorator';
 import { ApiBearerAuth, ApiBody, ApiConsumes, ApiCreatedResponse, ApiOkResponse, ApiParam } from '@nestjs/swagger';
 import UploadPhotoUserDto from './dto/upload-photo-user.dto';
 import { Express } from 'express';
